Evolução de mensagens de Logout

+++ src/components/Dropdowns/UserDropdown.js
@@ -2,8 +2,13 @@ import React from "react";
 import {useHistory} from 'react-router-dom';
 import { createPopper } from "@popperjs/core";
 import {FiPower} from 'react-icons/fi';
+import swal from 'sweetalert'
 
 function UserDropdown(e) {
+
+  const ongId   = localStorage.getItem('ongId');
+  const history = useHistory();
+
   // dropdown props
   const [dropdownPopoverShow, setDropdownPopoverShow] = React.useState(false);
   const btnDropdownRef = React.createRef();
@@ -18,10 +23,35 @@ function UserDropdown(e) {
     setDropdownPopoverShow(false);
   };
 
-  const history = useHistory();
+  
   function handleLogout(){
-    localStorage.clear();
-    history.push('/');
+
+    try {
+
+      if  (ongId)   {
+        localStorage.clear();
+        history.push('/');
+        swal({
+          title: "Usuário Desconectado!",
+          icon: "success",
+          button: "Ok!",
+        });    
+      } else {
+        swal({
+          title: "Nenhum Usuário Conectado !",
+          icon: "warning",
+          button: "Ok !",
+        });
+      }
+
+    } catch (err) {
+      swal({
+        title: "Algo deu errado !",
+        text: " VerTente novamente !",
+        icon: "error",
+        button: "Tentar Novamente !",
+      });
+    }
   }
 	
   return (
@@ -37,13 +67,12 @@ function UserDropdown(e) {
       >
         <div className="items-center flex">
           <span className="w-12 h-12 text-sm text-white bg-gray-300 inline-flex items-center justify-center rounded-full">
-          <button onClick={handleLogout} type="button">
-          <FiPower size={18} color="#e02041" />
-        </button>
+            <button onClick={handleLogout} type="button">
+              <FiPower size={18} color="#e02041" />
+            </button>
           </span>
         </div>
       </a>
-
     </>
   );
 };
